+++ Gulpfile.js
@@ -79,7 +79,6 @@ gulp.task('css', function(){
 	// Require.
 	var sass         = require('gulp-sass'),
 		sourcemaps   = require('gulp-sourcemaps'),
-		globbing     = require('gulp-css-globbing'),
 		autoprefixer = require('gulp-autoprefixer');
 
 	// Header template.
@@ -93,9 +92,6 @@ gulp.task('css', function(){
 
 	// Task.
 	return gulp.src(dist_dir + 'css/scss/build.scss')
-		.pipe(globbing({
-			extensions: ['.scss']
-		}))
         .pipe(sourcemaps.init())
         .pipe(sass({
 	        outputStyle: (is_production) ? 'compressed' : 'expanded',