+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 __X.X.X__:
 * Added ES6 Shim to dependencies.
+* Re-added missing gulp dependency
 
 __2.3.1__:
 * CSS debug mode disabled by default.